pyproject.toml
[tool.poetry]
name = "sentiment-data-sets"
version = "0.1.0"
description = ""
authors = ["Sean Dearnaley <SeanDearnaley@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
litellm = "^1.35.19"
openai = "^1.23.2"
python-dotenv = "^1.0.1"
pydantic = "^2.7.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
